Implemented task list widget compatability, updated UI
