--- studyspace/src/TaskListPage.css
@@ -1,10 +1,0 @@
-/* Organizing the page */
-.page-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    min-height: 100vh; /* Full viewport height */
-    background-size: cover;
-    /* background-position: center; */
-}
